PQClean Integration (#76)
* hashing benchmarks

* some more refactoring

* sample benchmarks

* sample benchmarks

* namespacing for pqclean

* namespacing in testvectors on host

* Integrate PQClean's AES into pqm4

* Fix path for host sources in Makefile

* Update mupq

* Fix bug in testvectors AES-128

* Add optimized AES-192/256 implementations.

Round-2-Kyber will use this.
Everything is put into separate sections such that the linker with
gc-sections will succesfully throw out an AES instance (or all of them)
whenever it is not used by a scheme.

* Update mupq

* fix incremental keccak implementation

* added benchmarks for ntruhps2048509 and sphincs-shake256-128f-simple

* make frodo scheme naming consistent with PQClean (breaks testvectors)

* rename frodo in benchmark results

+++ crypto_sign/dilithium/m4/api.h
@@ -1,5 +1,6 @@
 #ifndef API_H
 #define API_H
+#include <stddef.h>
 
 #ifndef MODE
 #define MODE 2
@@ -31,12 +32,12 @@
 
 int crypto_sign_keypair(unsigned char *pk, unsigned char *sk);
 
-int crypto_sign(unsigned char *sm, unsigned long long *smlen,
-                const unsigned char *msg, unsigned long long len, 
+int crypto_sign(unsigned char *sm, size_t *smlen,
+                const unsigned char *msg, size_t len, 
                 const unsigned char *sk);
 
-int crypto_sign_open(unsigned char *m, unsigned long long *mlen,
-                     const unsigned char *sm, unsigned long long smlen,
+int crypto_sign_open(unsigned char *m, size_t *mlen,
+                     const unsigned char *sm, size_t smlen,
                      const unsigned char *pk);
 
 #endif
